### 2018-05-29 #756 @whedon
[PRE REVIEW]: stacomiR : a common tool for monitoring fish migration
**Submitting author:** @MarionLegrandLogrami (<a href="http://orcid.org/0000-0002-4688-8303">Marion Legrand</a>)
**Repository:** <a href="https://github.com/MarionLegrandLogrami/stacomiR" target ="_blank">https://github.com/MarionLegrandLogrami/stacomiR</a>
**Version:** 0.5.3.1
