+++ paper.bib
@@ -40,21 +40,21 @@
 
 @book{poisson_eau_2011,
 	
-	edition = {Biotope -Muséum national d'Histoire naturelle},
+	edition = {Biotope -Museum national d'Histoire naturelle},
 	
-	series = {Inventaires \& {Biodiversité}},
+	series = {Inventaires et {Biodiversite}},
 	
-	title = {Les {Poissons} d'eau douce de {France} - {Biotope} éditions},
+	title = {Les {Poissons} d'eau douce de {France} - {Biotope} editions},
 	
 	url = {http://www.biotope-editions.com/index.php?article58/les-poissons-d-eau-douce-de-france},
 	
-	language = {Français},
+	language = {Francais},
 	urldate = {2018-05-17},
 	
 	author = {Keith, P. and Persat, H. and Feunteun, E and Allardi, J. (coords)},
 	
 	year = {2011},
 	
-	file = {Les Poissons d'eau douce de France - Biotope éditions:C\:\\Users\\marion.legrand\\Zotero\\storage\\XDWSPXIW\\index.html:text/html}
+	file = {Les Poissons d'eau douce de France - Biotope editions:C\:\\Users\\marion.legrand\\Zotero\\storage\\XDWSPXIW\\index.html:text/html}
 
   }
