### 2018-06-18 #756 @arfon
:wave: @steffilazerte @ben-williams - would you be willing to review this submission for JOSS?
