### 2018-06-25 #791 @whedon
[REVIEW]: stacomiR : a common tool for monitoring fish migration
**Submitting author:** @MarionLegrandLogrami (<a href="http://orcid.org/0000-0002-4688-8303">Marion Legrand</a>)
**Repository:** <a href="https://github.com/MarionLegrandLogrami/stacomiR" target ="_blank">https://github.com/MarionLegrandLogrami/stacomiR</a>
**Version:** 0.5.3.1

- [x] **Version:** Does the release version given match the GitHub release (0.5.3.1)?
