### 2018-06-25 #756 @arfon
@MarionLegrandLogrami @steffilazerte - please head over to #791 for the actual review.
