### 2018-06-26 419ec1b
Update paper.md
