### 2018-06-26 30be882
Update paper.md
