### 2018-06-26 31ea626
Update paper.md
