### 2018-06-28 1b9c469
Update paper.md
