### 2018-06-28 #791 @steffilazerte
The premis of `stacomiR` is to provide access to data base for monitoring fish migration in a user-friendly way. This is a great idea! I think it's really important to provide access to large biological data sets, especially in a way that is inclusive and which helps users who aren't familiar with R. The amount of work that has gone into this R package is also substantial and I'm eager to see it in action.

Unfortunately, I ran into several major issues when I tried to install and use `stacomiR`.

1) Lack of documentation
While the functions themselves are well documented, there are no tutorials, vignettes or websites with information for users just getting started. Even the documentation revealed by `?stacomiR-package` didn't provide any information for a new user, but merely directed the user to contact the author.

2) Installation
Dependencies in the DESCRIPTION don't list minimum versions. This may create problems for some users with older package versions. For example, I ran into a problem with the installation because I have Hmisc v4.0.3, but the package required Hmisc >= v4.1.0

3) Usage
Possibly because of both the above, once installed I struggled to use the package.

### Specific problems

I followed the instructions in the JOSS paper, but was unable to launch the interface with `stacomi()`, or with `stacomi(gr_interface = FALSE, login_window = FALSE, database_expected = FALSE)`.

I expected the first command to fail as I don't have the database (and it did) but the second command did nothing at all (no output, no errors), and I was at a loss of how to further proceed.

Following the rest of the example I ran into an error with the `calcule()` function:

```
> r_mig_mult <- calcule(r_mig_mult, silent = TRUE)
Loading required package: DBI
Error in postgresqlNewConnection(drv, ...) :
  RS-DBI driver: (could not connect test@localhost:5432 on dbname "test": could not connect to server: Connection refused
	Is the server running on host "localhost" (127.0.0.1) and accepting
	TCP/IP connections on port 5432?
)
Error in !dbPreExists : invalid argument type
```

The final lines in the JOSS paper gave me no output (presumably because the above function failed). I also struggled to understand what the functions were supposed to do.

### Other issues:

- Authorship: The package is authored by three people, but the JOSS paper only by two. Also the JOSS paper's first author seems to have made the fewest contributions (only judging by the number of Adds/Commits in the subversion repository). Perhaps the authors could comment on this.

- In the JOSS paper, the repository is listed as GitHub, but the GitHub repository isn't a complete R package (it's missing DESCRIPTION file, for one). It also looks as if the main source is located on R Forge. I would suggest either updating the GitHub repository to be instep with R Forge, or omitting the GitHub repository completely and just stick with R Forge.

- From what I understand of JOSS policies, there should actually be very little documentation in the JOSS paper. This paper should just be a summary of what the program is capable of, not information on how to use it.

- Which regions are included in this database? France? Europe? etc.?

### Verdict
In the end, I can't say I've actually reviewed the package, as I wasn't able to get it working. However, dealing with remote databases and interactive interfaces are always tricky. I'm still looking forward to giving it a shot and I'm more than happy to work with the authors to figure out what's going on.

For acceptance in JOSS I definitely need to finish my review (i.e. be able to test the package), and there also needs to be some solid documentation (perhaps a package vignette?) that walks new users through getting set up and addresses common problems or road blocks.

### My System Info

```
 version  R version 3.4.4 (2018-03-15)
 os       Ubuntu 16.04.4 LTS
 system   x86_64, linux-gnu
 ui       RStudio
 language en_CA:en
 collate  en_CA.UTF-8
 tz       America/Winnipeg
 date     2018-06-28
```
