### 2018-07-02 69e83b7
Update paper.md
