+++ paper.md
@@ -7,6 +7,8 @@ tags:
 authors:
  - name: Marion Legrand
    affiliation: "1, 2"
+ - name: Timothée Besse
+   affiliation: 1   
  - name: Cédric Briand
    affiliation: 3
 affiliations:
