### 2018-07-11 a381f41
Update paper.md
