+++ paper.md
@@ -7,10 +7,10 @@ tags:
 authors:
  - name: Marion Legrand
    affiliation: "1, 2"
- - name: Timothée Besse
-   affiliation: 1   
  - name: Cédric Briand
    affiliation: 3
+ - name: Timothée Besse
+   affiliation: 1
 affiliations:
  - name: Loire Grands Migrateurs (LOGRAMI), 5, avenue Buffon, Orléans, France
    index: 1
