### 2018-07-11 #791 @MarionLegrandLogrami
Dear ​Steffi LaZerte,

First of all, Cédric Briand and I wanted to warmly thank you for the time you spent to test our package. We found your remarks really useful and some will allow us to improve it​​.
I will take your remarks/questions one by one and in the order of your message to try to answer as best as possible.

**1. Lack of documentation**
Indeed, the documentation of the R package does not provide any help on how to use the package itself (link to the database, the installations to be made, etc.). To present in more details the package we wrote a vignette but we failed to compile and integrate it into the package. We did not understand why, but one of your remarks enlightened us on this point (I will come back to this later). Until we can integrate this vignette to our package you can already find it at this address : https://github.com/MarionLegrandLogrami/stacomiR/tree/master/vignettes.
In addition, we created a space dedicated to this project on "trac" system. This space has been created for users of the stacomi project. They can find different elements to help them to better use this tool. You can access to this space following this link : http://w3.eptb-vilaine.fr:8080/tracstacomi/wiki. You will need to connect to this space using user (utilisateur) : stacomi and password (mot de passe) : stacomi.
To give you some background on the fish data banking in France, you should know that there is no centralized system to organize this collection. The data are collected by various associations whose purpose is to follow fish on the long term (especially diadromous fish). Sometimes this is done by the Amateur Fishing Federations and the Migratory Associations (10 in France that work on diadromous fish) of the watershed concerned, pool the informations of the different Federations and make sure that the protocols are as homogeneous as possible. Sometimes it is the Migratory Associations themselves who collect the data (the most frequent case). Finally, there may be public institutions or research laboratories that monitor stations. To date, there is no structure to homogenize this collection or to organize the valuation of these data. It is on this basis that we decided to develop a R package to enable all data producers, whatever the structure, to improve the banking and the valuation of their data and if wanted to share data. Data ownership obviously remains that of producers and only producers can decide whether or not they want to make their data available and in what context. However, the R package is an excellent tool to show data producers the value of sharing information to compare data across multiple stations, multiple watersheds, and the group of Stacomi users that we created Cédric and I allows them to meet each other, share their package practice and make sharing information easier. Ultimately, we want to develop a package using Shiny and that would go further in the valuation and sharing of information.The stacomi project allows to make available a common tool and Cédric and I provide assistance to users concerning the installation of the tool, the difficulties of use, etc.

**2. Installation**
It's a good point, thank you for highlighting it. On rforge website, we indicated that to work the package needs to be run with r 3.5.1 in particular since some functions of the new version of Hmisc don't work anymore with previous version of R. But, it would be better to indicate the version of dependencies in the description as you propose. I wrote a ticket on the trac to remember this improvement for a new version of the package ([ticket number 238](http://w3.eptb-vilaine.fr:8080/tracstacomi/ticket/238)).

**3. Usage**
We hope that all clarifications gave in the two first point will help !

**4. Specific problems**
The problem you submitted to us regarding the impossibility of connecting to the "test" database gave us an interesting hint to understand why our vignette did not compile. For some calculations that we do in the package we have created a database named "test" that has been created for all users. Nevertheless, a person who does not have the stacomi installation obviously does not have a test database. We will work on a new solution to no longer need this database. I created a ticket on the trac corresponding to this improvement ([ticket number 239](http://w3.eptb-vilaine.fr:8080/tracstacomi/ticket/239)).
The commande line  stacomi() works only for users having a complete installation as you guessed. The command line  stacomi(gr_interface = FALSE, login_window = FALSE, database_expected = FALSE) allows you to use the package but only in r command line as the interface need a direct connexion to the database. The reason why the next command didn't work (r_mig_mult <- calcule(r_mig_mult, silent = TRUE)) is because we ask in the calcule method of this function to connect to the "test" database. This is a mistake and it is the reason I created a trac ticket ([239](http://w3.eptb-vilaine.fr:8080/tracstacomi/ticket/239)) to improve the next version of our package. To work around this problem immediately, is it possible to you to create a localhost with UTF8 encoding database under postgresql named "test" (landlord = postgres) and a test login role with user=test and password=test. This will allow the command line  r_mig_mult <- calcule(r_mig_mult, silent = TRUE)  to work.

**5. Other issues**

- Authorship: I added Timothée Besse in the author of the paper as he helps us to develop some parts of the package. For the order of the name in the paper, the reasons why we chose to put my name first even if I made much less commit than Cédric on the package are the following: while Cédric worked mainly on the development of the R package, I collected data on fish migration from more than 20 different structures in order to bank all the data in the postgresql database. This bankarization (more than 18 million of fish passage data) allowed the data producers to use the package's functionalities on all their data (including historical data) which was essential for the tool to become interesting for them. Thus, we divided the work mainly between Cédric and I to be the most effective. Of course we worked collaboratively which allowed everyone to follow the progress of the other and when necessary to give some help. On the other hand, I carry out a thesis on the modification of fish migration timing in France which required the use of the collected data (article being re-read by co-authors). It seemed to us more interesting to put my name first to value all this work in the context of my thesis (this choice is obviously shared by Cédric and I). To give you an idea of the work, here is a map of the monitoring stations whose data I banked. The oldest stations have a time series of more than 30 years

![stacomi_localisation_ac_conv_2018_03_23](https://user-images.githubusercontent.com/26621584/42285905-3ac40928-7fb1-11e8-9c3f-b7aa06a71351.png)

- Repository: Indeed, all the package is developped on rforge. But we thought that to submit a JOSS paper we had to have a github repository. That's why we copied / pasted the different files into the github directory created specifically to submit the paper to JOSS. As we didn't know how to change now the information for the directory, I update the github repository to add DESCRIPTION file.

- JOSS policies: Does this mean that you recommend to delete the different lines of R code we write on the paper to give some illustration on the possibilities of the package? [We saw for example this paper published on JOSS](https://joss.theoj.org/papers/87b007aadfc9f96dfff5fc001b34a2b0) which incorporated command lines of their package R and we thus thought that it was allowed. But if you think it would be better to delete them to give a better overview of what the package can do we are not opposed.

- Region included: at the moment the database is used all over France (thanks in large part to the banking work I have done for my thesis) and also in Spain. We have planned that the package can be used quite widely and that is why we have set up a system for translating the messages sent by the package in French and English for all messages and in Spanish for only a part (translation system realized with po.edit). To go further and promote the package in other countries, we should be able to put more resources on the translation, but if one day it becomes necessary, everything is planned so that we can do it.

We hope that these informations will be useful to you. Do not hesitate to come back to us if there are any questions or if you encounter difficulties.
Thank you again for the time invested, which was very helpful to us.
