### 2018-07-23 #791 @arfon
Thanks for the detailed review @steffilazerte. Responding to a couple of your comments here:

> This is good, but for an open source project, it might be better to have an open tracking system as well. At the very least there should be some information for users as to where they can go to submit issues. Consider using the "BugReports: " tag in the DESCRIPTION (for example see here)

This is actually a requirement for JOSS submissions (see https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements):

> ...the software associated with your submission must:
> - Have an issue tracker that is readable without registration

> Thank you for the clarification. I was sure there was a rational, but part of my job as a reviewer is checking up on this. @arfon, do you think it could be useful to have authors declare authorship contributions as part of the review application? It's not very common, but I know of at least some journals which require this.

We're considering this as an editorial board but at this point we don't have a policy for this.
