+++ README.md
@@ -5,4 +5,4 @@ migration, so that data from watershed are shared, and stocks exchanging between
 managed.
 The program is intended to be used by a "non experienced" R user, but all the R code automatically generated by the programm is shown to the user. Thus, it is possible to copy/paste the code and modify it (for example to change the preprogrammed colors or make more complicated changes). 
 The package is available from CRAN and a development version is available
-from R-Forge.
+from R-Forge. You can currently use this git as an open system to submit recommendations / report bugs.
