### 2018-09-24 #791 @MarionLegrandLogrami
Hi @arfon and @steffilazerte  - thanks for your message. A very short message just to tell you that we are incorporating the @steffilazerte's remarks. it's hard to tell you when we're done but we've made good progress. I will send you a message as soon as the package is ready and in accordance with the requests made
