### 2018-09-27 30bef2f
Add contributing file
help users to know how to contribute to the project
