+++ CONTRIBUTING.md
@@ -1,0 +1,21 @@
+# How to contribute to stacomiR
+
+You are welcome to contribute to the stacomiR project in different ways :
+
+## Issue tickets
+### Identify areas for future development ([open an Issue](https://github.com/MarionLegrandLogrami/stacomiR/issues))
+### Identify issues/bugs ([open an Issue](https://github.com/MarionLegrandLogrami/stacomiR/issues))
+For issue ticket :
+- Please ensure the bug or the future development was not already reported by searching on GitHub under [Issues](https://github.com/MarionLegrandLogrami/stacomiR/issues).
+- If a new issue is needed, please include a title and clear description, as much relevant information as possible, and a code sample or an executable test 
+case demonstrating the expected behavior that is not occurring.
+
+## Pull requests
+### Add functionality ([open a Pull Request](https://github.com/MarionLegrandLogrami/stacomiR/pulls))
+### Fix bugs ([open a Pull Request](https://github.com/MarionLegrandLogrami/stacomiR/pulls))
+For pull requests :
+- Please contribute via pull requests against the master branch.
+- When possible, please provide unit tests and documentation with code.
+
+## License
+This package is licensed under the GPL (version 2 or later) just like R itself. Your contributions should be licensed the same way.
