+++ Installation/README.md
@@ -4,7 +4,7 @@ The stacomiR package allows you to have access to your fish migratory database a
 
 For complete installation please be sure to have these softwares installed :
 - java 8
-- PostgreSQL
+- PostgreSQL (with pgAdmin - Open Source administration and development platform for PostgreSQL)
 - R >= 3.5.0
 
 You need also to:
@@ -13,15 +13,39 @@ You need also to:
 
 ## Create a database
 Install_bd_stacomi.sql (available in the [Installation folder](https://github.com/MarionLegrandLogrami/stacomiR/tree/master/Installation)) allows you to create an empty database. **You first need to decompress the file Install_bd_stacomi.zip**.
-You will find different schemas:
+Before restauring this file you will need to add different connections roles inside pgAdmin:
+- for iav
+```
+CREATE ROLE iav LOGIN PASSWORD 'iav'
+  NOSUPERUSER INHERIT NOCREATEDB NOCREATEROLE NOREPLICATION;
+```
+- for invite
+```
+CREATE ROLE invite LOGIN PASSWORD 'invite'
+  NOSUPERUSER INHERIT NOCREATEDB NOCREATEROLE NOREPLICATION;
+```
+- for user_1
+```
+CREATE ROLE user_1 LOGIN PASSWORD 'user_1'
+  NOSUPERUSER INHERIT NOCREATEDB NOCREATEROLE NOREPLICATION;
+```
+-for user_2
+```
+CREATE ROLE user_2 LOGIN PASSWORD 'user_2'
+  NOSUPERUSER INHERIT NOCREATEDB NOCREATEROLE NOREPLICATION;
+```
+You now need to restaure Install_bd_stacomi.sql. To do that open an MS DOS command and write (you first need to save the path to the Program\PostgreSQL\*your_version*\bin in Environment Variables):
+```
+psql -U *user* bd_contmig_nat<"path_to_the_Install_bd_stacomi.sql_file"
+```
+
+Now that your bd_contmig_nat database is restaured, you will find in your database different schemas (to look at your database, open pgAdmin and search for a database called bd_contmig_nat):
 - ref : with all the reference tables to understand codification used (we used the french standard 'SANDRE' for all codifications)
 - nat : for french users, this schema allow a national compilation of all users's schema in France. Please don't write inside this schema
 - iav : an example of a real user schema. In this schema you will find all data of the Etablissement Public Territorial du Bassin de la Vilaine.
 - user_1 & user_2 : two examples of empty user's schemas. Each user write in his schema. It is possible to share information between user. To do this you need to dump your schema (using MS DOS commande pg_dump or graphical interface). For example inside pgAdmin you can do right click on your schema and click "Backup". Then send this backup file to the user with whom you want to share data. He can restore your schema using MS DOS command (psql) or graphical interface (inside pgAdmin right click and "Restore")
 
-In pgAdmin, you need to create a connexion role with the same name as your schema (for example if I write inside schema called 'toto' I need to create a connexion role with name 'toto' and password 'toto' (or wathever password you choose)).
-
-If you want to add data in one of this empty user schema you will probably want to change the name. To do so, you just need to choose a  name for your schema (short name of your organization for example) and execute these lines :
+If you want to add data in one of this empty user schema you will probably want to change the name of one of the empty schema provided. To do so, you just need to choose a name for your schema (short name of your organization for example) and execute these lines :
 
 ```
 INSERT INTO ref.ts_organisme_org VALUES
@@ -29,6 +53,9 @@ INSERT INTO ref.ts_organisme_org VALUES
 
 ALTER SCHEMA user_1 RENAME TO name_of_your_schema;
 ```
+You also need to add new connection role:
[…]
