+++ R/report_mig_mult.R
@@ -188,9 +188,6 @@ setMethod("choice_c",signature=signature("report_mig_mult"),definition=function(
 #' @param silent Default FALSE, should messages be stopped
 #' @note The class does not handle escapement rates, though structurally those are present in the database. If you 
 #' want to use those you will have to do the calculation manually from the data in \code{report_mig_mult@data}.
-#' Note also that running the calcule method requires to have a database called test in postgres, and empty
-#' database in which all sqldf group by queries are run. The user and password for the test database are taken 
-#' from the calcmig.csv configuration file.
 #' @return report_mig_mult with a list in slot calcdata. For each dc one will find a list with the following elements
 #' \describe{
 #' \item{method}{In the case of instantaneous periods (video counting) the sum of daily values is done by the \link{fun_report_mig_mult} method and the value indicated in method is "sum".
@@ -208,7 +205,8 @@ setMethod("choice_c",signature=signature("report_mig_mult"),definition=function(
 #' @aliases calcule.report_mig_mult
 #' @export
 setMethod("calcule",signature=signature("report_mig_mult"),definition=function(object,negative=FALSE,silent=FALSE){ 
-	  # report_mig_mult<-r_mig_mult
+	  
+      # report_mig_mult<-r_mig_mult; negative=FALSE
 	  if (!silent) funout(gettext("Starting migration summary ... be patient\n",domain="R-stacomiR"))
 	  report_mig_mult<-object		
 	  debut=report_mig_mult@timestep@dateDebut
@@ -231,7 +229,7 @@ setMethod("calcule",signature=signature("report_mig_mult"),definition=function(o
 		  lestableaux[[stringr::str_c("dc_",dic)]][["method"]]<-"overlaps"
 		  contient_poids<-"poids"%in%datasub$type_de_quantite
 		  lestableaux[[stringr::str_c("dc_",dic)]][["contient_poids"]]<-contient_poids
-         
+          
 		  lestableaux[[stringr::str_c("dc_",dic)]][["negative"]]<-negative
 		  if (contient_poids){
 			coe<-report_mig_mult@coef_conversion[,c("coe_date_debut","coe_valeur_coefficient")]
@@ -413,7 +411,6 @@ h_report_mig_mult_graph=function(h=null,...){
 #' @aliases plot.report_mig_mult
 #' @export
 setMethod("plot",signature(x = "report_mig_mult", y = "missing"),definition=function(x, plot.type="standard",color=NULL, color_ope=NULL,silent=FALSE,...){ 
-	  #browser()
 	  #print("entering plot function")
 	  #report_mig_mult<-r_mig_mult;silent=FALSE
 	  report_mig_mult<-x
@@ -430,11 +427,11 @@ setMethod("plot",signature(x = "report_mig_mult", y = "missing"),definition=func
 		for (dcnum in 1:length(lesdc)){
 		  for (taxanum in 1:nrow(lestaxa)){
 			for (stagenum in 1:nrow(lesstage)){
-			  
-			  taxa=lestaxa[taxanum,"tax_nom_latin"]
-			  stage=lesstage[stagenum,"std_libelle"]
-			  dc=lesdc[dcnum]
-			  data<-report_mig_mult@calcdata[[stringr::str_c("dc_",dc)]][["data"]]
+			  #dcnum=1;taxnum=1;stagenum=1
+			  taxa <- lestaxa[taxanum,"tax_nom_latin"]
+			  stage <- lesstage[stagenum,"std_libelle"]
[…]
