+++ inst/config/test_vignette.R
@@ -1,3 +1,5 @@
- setwd("F:/workspace/stacomir/pkg/stacomir")
+ setwd("c:/workspace/stacomir/pkg/stacomir")
  devtools::build_vignettes()
+ # don't work
+# tools::compactPDF("/vignette/stacomi.pdf")
  
