+++ inst/examples/report_mig_interannual-example.R
@@ -18,8 +18,8 @@ stacomi(gr_interface=FALSE,
   assign("sch","pmp.",envir_stacomi)
   # Note in some cases you will want to change host and port setting
 #   sqldf.options<-get("sqldf.options",envir=envir_stacomi)  
-#   sqldf.options["sqldf.host"]<-"www.myhostname.com"
-#   sqldf.options["sqldf.port"]<-5433  
+#   sqldf.options["sqldf.RPostgreSQL.host"]<-"www.myhostname.com"
+#   sqldf.options["sqldf.RPostgreSQL.port"]<-5433  
 #   assign("sqldf.options",sqldf.options,envir_stacomi)  
   r_mig_interannual<-new("report_mig_interannual")
   r_mig_interannual<-choice_c(r_mig_interannual,
@@ -29,11 +29,14 @@ stacomi(gr_interface=FALSE,
 	  anneedebut="1990",
 	  anneefin="2015",
 	  silent=TRUE)
-  r_mig_interannual<-connect(r_mig_interannual,silent=TRUE)	
+  r_mig_interannual<-charge(r_mig_interannual)
+  r_mig_interannual<-connect(r_mig_interannual,check=TRUE)	
+  r_mig_interannual<-calcule(r_mig_interannual,silent=TRUE)	
 }	
 #############otherwise use this ######################
 # load the dataset generated by previous lines
 data("r_mig_interannual")
+
 #######################################################
 # the first plot is of little interest, it allows to see what data 
 # are available... simple lines
