+++ tests/testthat/test-02-report_mig.R
@@ -1,7 +1,7 @@
 context("report_mig")
 test_that("Test an instance of report_mig",{
-	  require(stacomiR)
-	  stacomi(gr_interface=FALSE,login_window=FALSE,database_expected=FALSE)
+      skip_if_not(stacomi_installed(),"skipping as the program is not installed on this computer")
+	  stacomi(gr_interface=FALSE,login_window=FALSE,database_expected=TRUE)
 	  # overriding user schema to point to iav
 	  baseODBC<-get("baseODBC",envir=envir_stacomi)
 	  baseODBC[c(2,3)]<-rep("iav",2)
@@ -24,8 +24,8 @@ test_that("Test an instance of report_mig",{
 
 
 test_that("Test an instance of report_mig, check that operations accross two years are split correcly",{
-	  require(stacomiR)
-	  stacomi(gr_interface=FALSE,login_window=FALSE,database_expected=FALSE)
+	  skip_if_not(stacomi_installed(),"skipping as the program is not installed on this computer")
+	  stacomi(gr_interface=FALSE,login_window=FALSE,database_expected=TRUE)
 	  # overriding user schema to point to iav
 	  baseODBC<-get("baseODBC",envir=envir_stacomi)
 	  baseODBC[c(2,3)]<-rep("iav",2)
@@ -54,8 +54,8 @@ test_that("Test an instance of report_mig, check that operations accross two yea
 	})
 
 test_that("Test another instance of report_mig, check that operations accross two years are split correcly",{
-	  require(stacomiR)
-	  stacomi(gr_interface=FALSE,login_window=FALSE,database_expected=FALSE)
+	  skip_if_not(stacomi_installed(),"skipping as the program is not installed on this computer")
+	  stacomi(gr_interface=FALSE,login_window=FALSE,database_expected=TRUE)
 	  # overriding user schema to point to iav
 	  baseODBC<-get("baseODBC",envir=envir_stacomi)
 	  baseODBC[c(2,3)]<-rep("iav",2)
@@ -84,9 +84,10 @@ test_that("Test another instance of report_mig, check that operations accross tw
 	})
 
 test_that("Test connect method",{
+      skip_if_not(stacomi_installed(),"skipping as the program is not installed on this computer")
 	  stacomi(gr_interface=FALSE,
 		  login_window=FALSE,
-		  database_expected=FALSE)	
+		  database_expected=TRUE)	
 	  # overriding user schema
 	  baseODBC<-get("baseODBC",envir=envir_stacomi)
 	  baseODBC[c(2,3)]<-rep("iav",2)
@@ -122,6 +123,7 @@ test_that("Test example 02_report_mig",
 
 test_that("Summary method works",
 	{
+      skip_if_not(stacomi_installed(),"skipping as the program is not installed on this computer")
 	  stacomi(gr_interface=FALSE,
 		  login_window=FALSE,
 		  database_expected=TRUE)	
@@ -140,6 +142,7 @@ test_that("Summary method works",
 
 test_that("Test writing an example to the database",
 	{
+      skip_if_not(stacomi_installed(),"skipping as the program is not installed on this computer")
 	  stacomi(gr_interface=FALSE,
 		  login_window=FALSE,
 		  database_expected=TRUE)	
@@ -158,6 +161,7 @@ test_that("Test writing an example to the database",
 
 test_that("Test that different sums are the same, for report_mig, report_mig_interannual, report_annual",
 	{
+      skip_if_not(stacomi_installed(),"skipping as the program is not installed on this computer")
 	  stacomi(gr_interface=FALSE,
[…]
