+++ man/charge_complement-ref_parqual-method.Rd
@@ -31,4 +31,3 @@ taxa_selectionne=2038
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
