+++ man/charge-ref_year-method.Rd
@@ -30,4 +30,3 @@ showMethods("charge")
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
