+++ man/calcule-report_mig_interannual-method.Rd
@@ -33,4 +33,3 @@ The class report_mig does not handle escapement rates nor
 \author{
 Marion Legrand
 }
-
