+++ man/charge-ref_stage-method.Rd
@@ -25,4 +25,3 @@ Loading method for ref_stage referential objects
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
