+++ man/charge.Rd
@@ -17,4 +17,3 @@ Generic method to load referentials
 \author{
 cedric.briand
 }
-
