+++ man/calcule.Rd
@@ -17,4 +17,3 @@ Generic method for calculations
 \author{
 cedric.briand
 }
-
