+++ man/choice_c-report_annual-method.Rd
@@ -35,4 +35,3 @@ command line interface for \link{report_annual-class}
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
