+++ man/choice_c-ref_taxa-method.Rd
@@ -28,4 +28,3 @@ choice_c(object=object,objectreport=objectreport,"Anguilla anguilla")
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
