+++ man/choice_c-report_sea_age-method.Rd
@@ -1,35 +1,35 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/report_mig_char.R
+% Please edit documentation in R/report_sea_age.R
 \docType{methods}
-\name{choice_c,report_mig_char-method}
-\alias{choice_c,report_mig_char-method}
-\alias{choice_c.report_mig_char}
-\title{command line interface for report_mig_char class}
+\name{choice_c,report_sea_age-method}
+\alias{choice_c,report_sea_age-method}
+\alias{choice_c.report_sea_age}
+\title{command line interface for report_sea_age class}
 \usage{
-\S4method{choice_c}{report_mig_char}(object, dc, taxa, stage,
-  parquan = NULL, parqual = NULL, horodatedebut, horodatefin,
-  echantillon = gettext("with", domain = "R-stacomiR"), silent = FALSE)
+\S4method{choice_c}{report_sea_age}(object, dc, taxa = 2220,
+  stage = c("5", "11", "BEC", "BER", "IND"), par = c("1786", "1785",
+  "C001", "A124"), horodatedebut, horodatefin, limit1hm, limit2hm,
+  silent = FALSE)
 }
 \arguments{
-\item{object}{An object of class \link{report_mig_char-class}}
+\item{object}{An object of class \link{report_sea_age-class}}
 
 \item{dc}{A numeric or integer, the code of the dc, coerced to integer,see \link{choice_c,ref_dc-method}}
 
-\item{taxa}{'2220=Salmo salar', can be a vector with several values
+\item{taxa}{'2220=Salmo salar',
 these should match the ref.tr_taxon_tax referential table in the stacomi database, see \link{choice_c,ref_taxa-method}}
 
-\item{stage}{The stages selected, can be a vector with several values}
+\item{stage}{'5','11','BEC','BER','IND'}
 
-\item{parquan}{Quantitative parameter}
-
-\item{parqual}{Qualitative parameter}
+\item{par}{Parameters chosen for the report are measured body size (1786), measured fork length (1785),video size (C001) and number of year at sea (A124)}
 
 \item{horodatedebut}{The starting date as a character, formats like \code{\%Y-\%m-\%d} or \code{\%d-\%m-\%Y} can be used as input}
 
 \item{horodatefin}{The finishing date of the report, for this class this will be used to calculate the number of daily steps.}
 
-\item{echantillon}{Default "with" can be "without" (values "avec" and "sans") are accepted, checking without modifies the query
-in the connect method so that subsamples are not allowed,}
+\item{limit1hm}{Size limit of a salmon for an one sea winter fish}
+
+\item{limit2hm}{Size limit of a salmon for a two sea winter fish}
 
 \item{silent}{Default FALSE, if TRUE the program should no display messages}
 }
@@ -39,7 +39,7 @@ The choice_c method fills in the data slot for classes \link{ref_dc-class}, \lin
 uses the choice_c methods of these object to select the data.
 }
 \description{
-command line interface for report_mig_char class
+command line interface for report_sea_age class
 }
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
