+++ man/connect-report_env-method.Rd
@@ -22,4 +22,3 @@ connect method for report_env class
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
