+++ man/funboxDC.Rd
@@ -18,4 +18,3 @@ function used for some lattice graph
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
 \keyword{internal}
-
