+++ man/connect-report_mig-method.Rd
@@ -19,4 +19,3 @@ report_mig with slot @data filled from the database
 \description{
 uses the report_mig_mult method
 }
-
