+++ man/connect-report_silver_eel-method.Rd
@@ -22,4 +22,3 @@ connect method for report_silver_eel
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
