+++ man/hexp.Rd
@@ -15,4 +15,3 @@ hexp(h, ...)
 Internal handler for reg, class \code{\link{report_ge_weight-class}}.
 }
 \keyword{internal}
-
