+++ man/h_report_sea_age.Rd
@@ -15,4 +15,3 @@ h_report_sea_age(h, ...)
 handler function used by the main interface
 }
 \keyword{internal}
-
