+++ man/getvalue.Rd
@@ -17,4 +17,3 @@ Generic method getvalue
 \author{
 cedric.briand
 }
-
