+++ man/interface_report_mig_char.Rd
@@ -13,4 +13,3 @@ interface for report_migPar class
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
 \keyword{internal}
-
