+++ man/ref_choice-class.Rd
@@ -17,16 +17,15 @@ radiobuttons
 
 \item{\code{selected}}{An \code{Integer}  the initial selected value (as an index), first=1 used in gradio}
 }}
+
 \section{Objects from the Class}{
  Objects can be created by calls of the form
 \code{new("ref_choice", listechoice=character() ,label=character()
 ,selected=integer())}.
 }
-\author{
-cedric.briand"at"eptb-vilaine.fr
-}
+
 \seealso{
-Other referential.objects: \code{\link{charge,ref_choice-method}},
+Other referential objects: \code{\link{charge,ref_choice-method}},
   \code{\link{ref_checkbox-class}},
   \code{\link{ref_coe-class}}, \code{\link{ref_dc-class}},
   \code{\link{ref_df-class}},
@@ -40,4 +39,7 @@ Other referential.objects: \code{\link{charge,ref_choice-method}},
   \code{\link{ref_taxa-class}},
   \code{\link{ref_year-class}}
 }
-
+\author{
+cedric.briand"at"eptb-vilaine.fr
+}
+\concept{referential objects}
