+++ man/ref_taxa-class.Rd
@@ -13,15 +13,14 @@ integrated into refreport objects.
 \describe{
 \item{\code{data}}{A \code{"data.frame"} of species available in the database}
 }}
+
 \section{Objects from the Class}{
  Objects can be created by calls of the form
 \code{new("ref_taxa", ...)}.
 }
-\author{
-cedric.briand"at"eptb-vilaine.fr
-}
+
 \seealso{
-Other referential.objects: \code{\link{charge,ref_choice-method}},
+Other referential objects: \code{\link{charge,ref_choice-method}},
   \code{\link{ref_checkbox-class}},
   \code{\link{ref_choice-class}},
   \code{\link{ref_coe-class}}, \code{\link{ref_dc-class}},
@@ -35,4 +34,7 @@ Other referential.objects: \code{\link{charge,ref_choice-method}},
   \code{\link{ref_stage-class}},
   \code{\link{ref_year-class}}
 }
-
+\author{
+cedric.briand"at"eptb-vilaine.fr
+}
+\concept{referential objects}
