+++ man/ref_period-class.Rd
@@ -12,6 +12,7 @@ ref_period referential class to choose a period
 \describe{
 \item{\code{data="data.frame"}}{providing correspondance between period and their English names}
 }}
+
 \note{
 pgval are used by seq.POSIXt
 }
@@ -19,8 +20,9 @@ pgval are used by seq.POSIXt
  Objects can be created by calls of the form
 \code{new("ref_period", ...)}.
 }
+
 \seealso{
-Other referential.objects: \code{\link{charge,ref_choice-method}},
+Other referential objects: \code{\link{charge,ref_choice-method}},
   \code{\link{ref_checkbox-class}},
   \code{\link{ref_choice-class}},
   \code{\link{ref_coe-class}}, \code{\link{ref_dc-class}},
@@ -34,5 +36,5 @@ Other referential.objects: \code{\link{charge,ref_choice-method}},
   \code{\link{ref_taxa-class}},
   \code{\link{ref_year-class}}
 }
+\concept{referential objects}
 \keyword{classes}
-
