+++ man/summary-report_silver_eel-method.Rd
@@ -21,4 +21,3 @@ summary for report_silver_eel
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
