+++ man/write_database-report_sea_age-method.Rd
@@ -22,4 +22,3 @@ defined in calcmig.csv
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
