+++ man/summary-report_df-method.Rd
@@ -21,4 +21,3 @@ summary for report_df, write csv and html output, and prints summary statistics
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
