+++ man/xtable-report_annual-method.Rd
@@ -39,4 +39,3 @@ xtable function for \link{report_annual-class}
 create an xtable objet but also assigns an add.to.column argument in envir_stacomi,
 for later use by the print.xtable method.
 }
-
