+++ man/summary-report_mig_interannual-method.Rd
@@ -30,4 +30,3 @@ historical dataset.
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
