+++ man/supprime-ref_coe-method.Rd
@@ -25,4 +25,3 @@ supprime method for "ref_coe" class
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
