+++ man/report_mig_interannual-class.Rd
@@ -2,8 +2,8 @@
 % Please edit documentation in R/report_mig_interannual.R
 \docType{class}
 \name{report_mig_interannual-class}
-\alias{report_mig_interannual}
 \alias{report_mig_interannual-class}
+\alias{report_mig_interannual}
 \title{Class "report_mig_interannual"}
 \description{
 When daily report are written in the t_reportjournalier_bjo table by the 
@@ -35,6 +35,7 @@ ref_year allows to choose last year of the report}
 
 \item{\code{calcdata}}{A \code{list} of calculated data, filled in by the calcule method}
 }}
+
 \examples{
 require(stacomiR)
 # launching stacomi without selecting the scheme or interface
@@ -54,6 +55,11 @@ stacomi(gr_interface=FALSE,
   assign("baseODBC",baseODBC,envir_stacomi)
   sch<-get("sch",envir=envir_stacomi)
   assign("sch","pmp.",envir_stacomi)
+  # Note in some cases you will want to change host and port setting
+#   sqldf.options<-get("sqldf.options",envir=envir_stacomi)  
+#   sqldf.options["sqldf.RPostgreSQL.host"]<-"www.myhostname.com"
+#   sqldf.options["sqldf.RPostgreSQL.port"]<-5433  
+#   assign("sqldf.options",sqldf.options,envir_stacomi)  
   r_mig_interannual<-new("report_mig_interannual")
   r_mig_interannual<-choice_c(r_mig_interannual,
 	  dc=c(16),
@@ -62,11 +68,14 @@ stacomi(gr_interface=FALSE,
 	  anneedebut="1990",
 	  anneefin="2015",
 	  silent=TRUE)
-  r_mig_interannual<-connect(r_mig_interannual,silent=TRUE)	
+  r_mig_interannual<-charge(r_mig_interannual)
+  r_mig_interannual<-connect(r_mig_interannual,check=TRUE)	
+  r_mig_interannual<-calcule(r_mig_interannual,silent=TRUE)	
 }	
 #############otherwise use this ######################
 # load the dataset generated by previous lines
 data("r_mig_interannual")
+
 #######################################################
 # the first plot is of little interest, it allows to see what data 
 # are available... simple lines
@@ -203,11 +212,8 @@ plot(r_mig_interannual,plot.type="density",silent=TRUE)
 
 
 }
-\author{
-Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
-}
 \seealso{
-Other report.Objects: \code{\link{report_annual-class}},
+Other report Objects: \code{\link{report_annual-class}},
   \code{\link{report_dc-class}},
   \code{\link{report_df-class}},
   \code{\link{report_env-class}},
@@ -221,5 +227,8 @@ Other report.Objects: \code{\link{report_annual-class}},
   \code{\link{report_silver_eel-class}},
   \code{\link{report_species-class}}
 }
+\author{
+Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
+}
+\concept{report Objects}
 \keyword{classes}
-
