+++ man/stacomi.Rd
@@ -56,4 +56,3 @@ stacomi(gr_interface=FALSE,login_window=FALSE,database_expected=FALSE)
 \author{
 Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
 }
-
