+++ Installation/README.md
@@ -34,7 +34,7 @@ CREATE ROLE user_1 LOGIN PASSWORD 'user_1'
 CREATE ROLE user_2 LOGIN PASSWORD 'user_2'
   NOSUPERUSER INHERIT NOCREATEDB NOCREATEROLE NOREPLICATION;
 ```
-You now need to restaure Install_bd_stacomi.sql. To do that open an MS DOS command and write (you first need to save the path to the Program\PostgreSQL\*your_version*\bin in Environment Variables):
+You now need to restaure Install_bd_stacomi.sql. To do that open an MS DOS command and write (you first need to save the path to the Program\PostgreSQL\your_version\bin in Environment Variables):
 ```
 psql -U user <"path_to_the_Install_bd_stacomi.sql_file"
 ```
