### 2018-10-16 1a8a2d1
Update paper.md
