### 2018-10-16 843f251
Update paper.md
