### 2018-10-18 c2d6c7f
Update paper.md
