### 2018-10-18 #791 @MarionLegrandLogrami
Hi @arfon and @steffilazerte - I'm happy to inform you that we finished to update the stacomiR package. The new version is available on CRAN and rforge (we updated the version provided on github as well). We made the main following modifications:

- remove all call to test database,

- incorporate the vignette inside the package,

- add [detailed installation instructions on github](https://github.com/MarionLegrandLogrami/stacomiR/blob/master/Installation/README.md) for new user,

- provide a database with one example schema (with data) and one empty schema (for people who want to use the database to organize their data),

- add [contribution file](https://github.com/MarionLegrandLogrami/stacomiR/blob/master/CONTRIBUTING.md) and a BugReports part inside the [description file](https://github.com/MarionLegrandLogrami/stacomiR/blob/master/DESCRIPTION),

- update the paper we submitted to add some context informations and add a link to the vignette.

We hope that this new version will meet all your requirements.
Again we wanted to thank @steffilazerte for all the really helpful comments that helped us to improve our package.
