### 2018-10-30 #791 @steffilazerte
@MarionLegrandLogrami and coauthors,

I'm happy to say that with the updates to the `stacomiR` package I've been able to run through the examples based on the built-in example data sets!

The vignette is a wealth of information and a valuable addition. By running through the example data, the plot/report documentation and checking back with the vignette's section on Data structure, I was able to gain a much better understanding of the kind of data being summarized (which is impressive, given how little I know about fish monitoring).

I was able to snoop through most of the documentation and create many different graphical summaries of the included data sets. I explored figures and summaries for these data sets and, once again, I am seriously impressed by the variety of analyses included.

Including instructions for installing/restoring the database on GitHub is a great idea, and I like that you've linked to them in the JOSS manuscript as well.

The downside is that I am still unable to setup the database, and I haven't been able to launch the user interface.

I've also realized that your setup is designed for Windows (whereas I am on Linux). For example, I will never be able to have the location `c:\users\my_session\Documents\CalcmigData` on my computer, as Linux file paths do not have drive locations (`c:\`) (and neither do Macs).

Below are the error messages I got when attempting to restore the database. I suspect that the `locale` error is related to the Operating system differences, but I am not sure if the `database "bd_contmig_nat" does not exist` error is the result of the `locale` error or something else.

```
postgres@steffi-desktop:~$ psql postgres < install_bd_contmig_nat.sql
SET
SET
SET
SET
SET
 set_config
------------

(1 row)

SET
SET
SET
ERROR:  invalid locale name: "French_France.1252"
ERROR:  database "bd_contmig_nat" does not exist
\connect: FATAL:  database "bd_contmig_nat" does not exist
postgres@steffi-desktop:~$
```

### Suggestions

- In the installation instructions, add Windows as a system requirement

- In my quest to access the database, I solved at least one error by updating my PostgreSQL to version 9.6, you may want to add this version requirement to the installation instructions.

- Be **very** explicit about the necessity of having a database in the vignette. I would suggest right at the top, under the "Usage" section, that you state that the following two commands will only work once you've installed (or created) a database. Then link to the installation instructions on GitHub.

- In the vignette example with silver eels and Durif stages, should it be plot.type = 1? or plot.type = 2? (I'm definitely not an expert, but expected to see the word 'Durif' somewhere on the figure)

### Conclusions

@arfon, I have evaluated everything I can here. The package authors have addressed my comments and concerns and I have run through all the JOSS review check boxes, with the exception of fully testing the functionality (i.e. accessing the database in order to create reports of novel data subsets and the graphical user interface). Admittedly, these aspects are a pretty big part of this R package.

I would suggest either accepting the revisions as sufficient, or finding another, windows-based reviewer to finish the review. My apologies for not being able to take this to the end. I hope this hasn't been too inconvenient for everyone involved!
