### 2018-11-04 #791 @arfon
> @arfon, I have evaluated everything I can here. The package authors have addressed my comments and concerns and I have run through all the JOSS review check boxes, with the exception of fully testing the functionality (i.e. accessing the database in order to create reports of novel data subsets and the graphical user interface). Admittedly, these aspects are a pretty big part of this R package.

Thanks for your excellent review @steffilazerte and your suggestion here. I think we should seek an additional reviewer to help verify these aspects of the package.

@MarionLegrandLogrami - do you know of anyone who might be able to help with this? @karthik - do you know of any Windows/R users that might be able to help?
