### 2018-11-15 #791 @MarionLegrandLogrami
@steffilazerte, my colleagues and I wanted to thank you once again for all the work done and the help provided. Thanks to you we made lot of progress ! For the error concerning the database

> database "bd_contmig_nat" does not exist
I think this is not related to the operating system. I saved the database using the pg_dump command with the -C option to insert in the dump file the command line needed to create the database but unfortunately (I do not understand why and I did not see this problem before reading your message) the psql command fails to restore the database without the user manually creating the database using the following SQL command line:
> CREATE DATABASE bd_contmig_nat WITH ENCODING = 'UTF8';
I will add this command line in the github instruction file for the installation of the database.

Thanks again for all your work and all your remarks and hints.
