+++ Installation/Instruction Installation Français.md
@@ -11,55 +11,84 @@ Vous devrez aussi :
 - Créer un lien ODBC pour votre base de données poissons migrateurs (ajouter une source de type "PostgreSQL unicode" - si vous n'avez pas le driver PostgreSQL, utilisez stack builder pour installer le driver)
 - Ajouter un dossier CalcmigData sur votre ordinateur (par exemple c:\users\my_session\Documents\CalcmigData)
 
-# Créer une base de données
+## Créer une base de données
 
 Avant tout chose, créez la base de données dans pgAdmin via une console SQL en exécutant le code suivant :
+```
 CREATE DATABASE bd_contmig_nat WITH ENCODING = 'UTF8';
+```
 Puis, ajoutez les différents rôles de connexion dans pgAdmin : 
 pour le rôle iav
+```
 CREATE ROLE iav LOGIN PASSWORD 'iav'
   NOSUPERUSER INHERIT NOCREATEDB NOCREATEROLE NOREPLICATION;
+```
 pour le rôle invite
+```
 CREATE ROLE invite LOGIN PASSWORD 'invite'
   NOSUPERUSER INHERIT NOCREATEDB NOCREATEROLE NOREPLICATION;
+```
 pour le rôle user_1
+```
 CREATE ROLE user_1 LOGIN PASSWORD 'user_1'
   NOSUPERUSER INHERIT NOCREATEDB NOCREATEROLE NOREPLICATION;
+```
 
-Téléchargez maintenant le fichier Install_bd_stacomi.zip (disponible dans le dossier Installation) et décompressez-le . 
+Téléchargez maintenant le fichier Install_bd_stacomi.zip (disponible dans le [dossier Installation](https://github.com/MarionLegrandLogrami/stacomiR/tree/master/Installation)) et décompressez-le. 
 Vous pouvez maintenant restaurer le fichier Install_bd_stacomi.sql. Pour cela, ouvrez une commande MS DOS (vous devez premièrement sauvegarder le chemin vers Program\PostgreSQL\your_version\bin dans les variables d’environnement) :
+```
 psql -U user  bd_contmig_nat<"path_to_the_Install_bd_stacomi.sql_file"
-Remplacez user par le nom que vous utilisez pour vous connecter dans pgAdmin. Cette commande fonctionne uniquement si votre base de données est installée sur votre serveur local avec le port par défaut (5432). Si votre base de données est installée sur un serveur distant, vous devez spécifier le nom du serveur et le port en ajoutant le paramètre -h nom du serveur ou l'adresse et -p le port.
+```
+* Avec user par le nom que vous utilisez pour vous connecter dans pgAdmin.*
+Cette commande fonctionne uniquement si votre base de données est installée sur votre serveur local avec le port par défaut (5432). Si votre base de données est installée sur un serveur distant, vous devez spécifier le nom du serveur et le port en ajoutant le paramètre -h *nom du serveur ou l'adresse* et -p *le port*.
 Votre base de données bd_contmig_nat est maintenant restaurée, vous trouverez dans celle-ci différents schémas (pour regarder votre base de données, ouvrez pgAdmin et cherchez la base de données appelée bd_contmig_nat):
-ref : contient toutes les tables de référence pour comprendre la codification utilisée (nous avons utilisé le standard "SANDRE" pour l'ensemble des codifications)
-nat : pour les utilisateurs français, ce schéma permet une compilation de l'ensemble des schémas des utilisateurs en France. Merci de ne pas modifer ce schéma.
-iav : un exemple d'un schéma utilisateur. Dans ce schéma vous trouverez toutes les données de l'Etablissement Public Territorial du Bassin de la Vilaine.
[…]
