### 2018-12-27 #791 @boshek
Rather than test on Windows which so far I have unable to accomplish I'd suggest two steps that would make this package more usable:
- Move the workflow of steps from a click based approach to R. For example, something like this:
> You need to download the stacomi.zip file, unzip the file and put the stacomi folder at the root of your c:\Program Files repository. You will find 6 files inside the stacomi folder

could likely be turned into a function using some combination of `download.file` and `unzip`.  Another example of this is the sql used to create the database. I believe some of this can directly be run from R (as well as the above download/unzip workflow described above).

- More importantly though I think this package should very strongly consider adapting code so that it can be run on multiple platforms. If broad adoption is a goal here, restricting to Windows seems like a mistake. Fortunately, the `rappdirs` package provides a very nice option of platform-independent locations to download data (i.e. the content of stacomi.zip). So that is you'd replace something like this:

```
filecsv<-"C:/Program Files/stacomi/calcmig.csv"
```

with this:

```
> filecsv <- file.path(rappdirs::site_data_dir("stacomiR"), "calcmig.csv")
> filecsv
[1] "C:\\PROGRA~3\\stacomiR\\stacomiR/calcmig.csv"
```
