### 2019-01-15 #791 @MarionLegrandLogrami
Hi @boshek . Thanks a lot for your comments.
First of all, I wanted to tell you that we are looking at your proposal to simplify the installation process, but we need some time to do it. I will come back to you when it's over
Regarding the use of multiple platforms, the stacomiR package [is already working on linux](https://r-forge.r-project.org/R/?group_id=1019). We have not checked the use of the package under Linux because we are all working on Windows, but we will consider configuring a virtual machine with Linux to test it. At this moment, all the users of this package are using windows operating system, except one that is using OS Mac. All these users are not very familiar with R (most have never used R) no more than with databases and ODBC link creation. For all these users we made all the installation for them. But we agree that trying to find a way to simplify this installation would be a good thing.

To respond more specifically to some of your comments :

> You also need to add new connection role: [...]

You need to do this **only** if you want to use the database (= add data) and you want to change the name of the user's schema (I will try to explain that better on github). To simply test the package you don't need to do that as you have the IAV schema with all data to test the package.

It's the same for the Updating_grant.zip file. You must run the file only if you want to change the schema username (you just need to open the SQL script file in pgAdmin and execute the code).

Regarding the error you get:

> > stacomi()
Error in funout(msg3, arret = TRUE) : connection impossible
In addition: Warning messages:
1: In RODBC::odbcDriverConnect("DSN=bd_contmig_nat;UID=iav;PWD=iav",  :
  [RODBC] ERROR: state IM002, code 0, message [Microsoft][ODBC Driver Manager] Data source name not found and no default driver specified
2: In RODBC::odbcDriverConnect("DSN=bd_contmig_nat;UID=iav;PWD=iav",  :
  ODBC connection failed

- Do you have a PostgreSQL ODBC driver ? If you have an R 32 bits you need a PostgreSQL ODBC driver 32 bits (if R is 64 bits you need a 64 bits ODBC driver).

- Did you configure a PostgreSQL ODBC link (Unicode), with the name of the database, the server and port, etc. ?

I'll be back to you shortly to inform you of our progress on the improvements you have submitted to us.
