+++ README.md
@@ -1,10 +1,13 @@
+
+
 # Welcome to the stacomiR project !
 
-It is a part of the 'STACOMI' open source project developed in France by the French Agency for Biodiversity (AFB) institute to centralize data obtained by fish pass monitoring.  The objective of the stacomi project is to provide a common database for people monitoring ﬁsh
-migration, so that data from watershed are shared, and stocks exchanging between diﬀerent basins are better
-managed.
+It is a part of the 'STACOMI' open source project developed in France by the French Agency for Biodiversity (AFB) institute, to centralize data obtained by fish pass monitoring.  The objective of the stacomi project is to provide a common database for people monitoring ﬁsh migration, so that data from watershed are shared, and stocks exchanging between diﬀerent basins are better managed.
+
 The program is intended to be used by a "non experienced" R user, but all the R code automatically generated by the programm is shown to the user. Thus, it is possible to copy/paste the code and modify it (for example to change the preprogrammed colors or make more complicated changes). 
-The package is available from CRAN and a development version is available
-from R-Forge. You can currently use this git as an open system to [submit recommendations / report bugs](https://github.com/MarionLegrandLogrami/stacomiR/issues).
+
+The package is available from CRAN and a development version is available from R-Forge. You can currently use this git as an open system to [submit recommendations / report bugs](https://github.com/MarionLegrandLogrami/stacomiR/issues).
 
 You can find all the [instructions needed for a full installation here](https://github.com/MarionLegrandLogrami/stacomiR/tree/master/Installation).
+
+And if you want to use it for your own structure, you will find some help on it's usage on the [How To page](https://github.com/MarionLegrandLogrami/stacomiR/tree/master/HOWTO.md).
