### 2019-02-24 6516c45
Linux & Windows install scripts
