+++ Installation/install_stacomi.sh
@@ -1,0 +1,309 @@
+#!/bin/bash
+#
+# Install script for Stacomi database on Linux
+# (c)2019 LOGRAMI      - Marion LEGRAND
+# (c)2019 EPTB Vilaine - Cédric BRILLANT
+# 
+# Return codes:
+#	 0   no error
+#    1   Missing dependency
+#    2   Error while processing
+#  255   Display usage
+#
+
+# ------------- Do not change anything above this line -------------
+
+# Defaults files
+_ROOT=$(cd $(dirname "${BASH_SOURCE[0]}") && pwd)
+ZIP_ARCH_NAME=install_bd_contmig_nat
+ZIP_ARCH=data/${ZIP_ARCH_NAME}.zip
+
+# Default variables
+DB_NAME=bd_contmig_nat
+SQL_SUDO=postgres
+SQL_USER=postgres
+SQL_PASS=postgres
+SQL_HOST=localhost
+SQL_PORT=5432
+REMOTE=0
+
+# Return codes
+RET_OK=0
+RET_MISS=1
+RET_ERR=2
+RET_USAGE=255
+RET=$RET_OK
+
+# Colors
+WHITE="\E[0m"
+BOLD="\E[1m"
+RED="$WHITE\E[31m"
+GREEN="$WHITE\E[32m"
+ORANGE="$WHITE\E[33m"
+BLUE="$WHITE\E[36m"
+
+# Executables
+PROG_UNZIP=unzip
+PROG_PSQL=psql
+
+
+# Supported OS
+#-------------
+OS_UPD=0
+OS=
+PACKAGE_MANAGER=
+[ -f /etc/redhat-release ] && OS=redhat && PACKAGE_MANAGER="yum"
+[ -f /etc/debian_version ] && OS=debian && PACKAGE_MANAGER="apt-get"
+
+
+#---------------------------------------------------------
+# Test dependency
+#	$1	program to test (empty = search for installed package name)
+#	$1	Description
+#	$2	Debian package name
+#	$3	RedHat package name
+#---------------------------------------------------------
+test_dependency() {
+	local prog=$1
+	local descr=$2
+	local package_debian=$3
+	local package_redhat=$4
+	echo -en "    [${BLUE}*${WHITE}] $descr"
+	
+	# What is the package name for the distribution?
+	local package=$(eval echo \$$"package_$OS")
+	
+	# Program given? We test if we find it in the PATH
+	local found=
+	if [ -n "$prog" ]; then
+		found=$(which $prog)
+	# If not, we check if the corresponding package is installed
+	# (/!\ It could have been compiled by hand, and so not being detected here...)
+	else
+		if [ "$OS" = "debian" ]; then
+			found=$(dpkg-query -W $package 2>/dev/null)
+		elif [ "$OS" = "redhat" ]; then
+			found=$(rpm -qa | grep "^${package}-" 2>/dev/null)
+		fi
+	fi
+	# found?
+	if [ -n "$found" ]; then
+		echo -e "[${GREEN}found${WHITE}] ($found)"
+	# Not found?
+	else
+		echo -e "[${RED}not found${WHITE}]"
+		# If we don't know this OS, we can't do anything
+		if [ -z "$OS" ]; then
+			echo "        --> Please install it manually (I can't do it by myself, I don't"
+			echo "            know how to do it with your Linux distrib, sorry...)"
+			exit $RET_MISS
+		fi
+		# If it is ok, we propose to install it automatically
+		echo "          I can install it automatically for you with $PACKAGE_MANAGER"
+		local choice=
+		read -p "          Install this dependency? (Y/n): " choice
[…]
