+++ Installation/README.md
@@ -1,104 +1,190 @@
-The stacomiR package allows you to have access to your fish migratory database and to plot different pre-programmed figures. With this package, we developped a java program to allow non-sql speaker to update or insert new data inside their postgreSQL fish migratory database.
+The stacomiR package allows you to have access to your fish migratory database and to plot different pre-programmed figures. With this package, we developed a java program to allow non-sql speaker to update or insert new data inside their postgreSQL fish migratory database.
 
-# Installation instructions
 
-For complete installation please be sure to have these softwares installed :
-- java 8
-- PostgreSQL >=9.5 (with pgAdmin - Open Source administration and development platform for PostgreSQL)
-- R >= 3.5.0
+# Table of contents
+1. [Requirement ](#dependencies)
+2. [Database automated installation](#auto-install)
+2.1. [On Linux](#linux-install)
+2.2. [On Windows](#win-install)
+3. [Manual installation](#manual-install)
+3.1. [Create a database](#manual-install-db)
+3.2. [Populating it](#manual-install-db-fill)
+3.3. [Create an ODBC connector](#manual-install-odbc)
+4. [Final words](#final)
 
-You will need also to:
-- create ODBC link for your fish migratory database (add a source of type "PostgreSQL unicode" - if you don't have a PostgreSQL ODBC driver please use stack builder to install the driver first or download the driver on the [postgresql website](https://www.postgresql.org/ftp/odbc/versions/))
-- (optional - if you want to choose where the folder must be created) add a **CalcmigData** folder on your computer (for example c:\users\my_session\Documents\CalcmigData)
 
-## Create a database
-Open pgAdmin and in a SQL script console type:
+# 1 Requirement <a name="dependencies"></a>
+
+For minimal installation, the following softwares must already be installed on your computer:
+- PostgreSQL ≥ 9.5 (with pgAdmin - Open Source administration and development platform for PostgreSQL)
+- PostgreSQL ODBC driver (Link: [PostgreSQL website](https://www.postgresql.org/ftp/odbc/versions/))
+- R ≥ 3.5.0
+
+For complete installation, you  also need:
+- java ≥ 8
+
+This wiki will help you to prepare your environment, by:
+- creating the test database,
+- populating it,
+- create an ODBC link to this database.
+
+############## Jérémie : à virer ? ################
+You will also need to:
+- create an ODBC link for your fish migratory database (add a source of type "PostgreSQL unicode". If you don't have a PostgreSQL ODBC driver please use stack builder to install it first, or download it from the [PostgreSQL website](https://www.postgresql.org/ftp/odbc/versions/))
+- [optional - if you want to choose where the folder must be created] add a **CalcmigData** folder on your computer (for example `c:\users\<my_session>\Documents\CalcmigData` on Windows, or `~/CalcmigData` on Linux)
+
+############## fin ################
+
+# 2 Database automated installation <a name="auto-install"></a>
+We made some scripts to help you creating the database, filling it and init the ODBC link.
+They have been tested on:
+- Microsoft Windows 10
+- Linux Debian 9 (should work with Ubuntu)
+- Linux CentOS 7 (should work with RedHat)
+
+
+## 2.1 On Linux host <a name="linux-install"></a>
+The script is:
+`install_stacomi.sh`
+
+Calling it with `-h` will display the possible options:
 ```
+/bin/batch install_stacomi.sh -h
+Install Stacomi database on Linux
+*** This script must be executed with administrative rights! ***
+
+Usages:
+
+  Server on same host than this script (default behavior):
+  --------------------------------------------------------
+  install_stacomi.sh [--sudo=account]
+
+  Options :
+    -h, --help                 This help
[…]
