### 2019-02-26 5ffffc4
Update install_stacomi.sh
