### 2019-02-28 ac30360
delete stacomiR package version as it was only a copy of the CRAN version (manual copy not updated). Easier for an user to directly download the package on CRAN (or on rforge for the development version)
