+++ R/ref_horodate.R
@@ -1,218 +1,0 @@
-#' Validity check for ref_horodate
-#' 
-#' @param object A ref_horodate object
-#' @author Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
-#' @keywords internal
-validity_ref_horodate=function(object)
-{
-  rep1= class(object@horodate)[2]=="POSIXt"
-  
-  return(ifelse(rep1,TRUE,FALSE))
-}
-
-
-#' Class ref_horodate
-#' 
-#' choice of date with method to show current and previous year
-#' 
-#' 
-#' @slot horodate a "POSIXt"
-#' @section Objects from the Class: Objects can be created by calls of the form
-#' \code{new("ref_horodate", \dots{})}.
-#' @author Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
-#' @family referential objects
-setClass(Class="ref_horodate",representation=
-		representation(horodate="POSIXt"),
-	validity=validity_ref_horodate,
-	prototype=prototype(horodate=Hmisc::roundPOSIXt(Sys.time(),"years")))
-# date= new("Horodate")
-#retourne la date en format character
-setGeneric("getref_horodate",def=function(object,...) standardGeneric("getref_horodate"))
-setMethod("getref_horodate",signature=signature("ref_horodate"),definition=function(object){
-	  return ( strftime(as.POSIXlt(object@horodate),format="%Y-%m-%d %H:%M:%S") )
-	})
-
-#Fixe la date de debut e partir d'un champ charactere de type "%Y-%m-%d %H:%M:%S"
-setGeneric("setref_horodate",def=function(object,...) standardGeneric("setref_horodate"))
-
-
-
-
-#' Method to set the Horodate
-#' @param object An object of class \link{ref_horodate-class}
-#' @param string A string representing an horodate in the format "\%Y-\%m-\%d \%H:\%M:\%S"
-#' @return An Object of class "ref_horodate" 
-#' @author cedric.briand
-#' @keywords internal
-setMethod("setref_horodate",signature=signature("ref_horodate"),definition=function(object,string){
-	  object@horodate=strptime(string,format="%Y-%m-%d %H:%M:%S")
-	  return(object) 
-	})
-# retourne l'annee d'avant l'annee en cours
-
-
-
-
-
-
-#' Graphical interface
-#' @param object An object of class \link{ref_horodate-class}
-#' @param label Label for the gframe
-#' @param nomassign The name assigned in environment envir_stacomi
-#' @param funoutlabel, text displayed by the interface
-#' @param decal Default 0, number of years to shift forward or backward 
-#' @keywords internal
-#' @return Selects the date in the graphical interface, and assigns an object of class POSIXt with name nomassign in envir_stacomi
-setMethod("choice",signature=signature("ref_horodate"),definition=function(object,
-		label="date",
-		nomassign="horodate",
-		funoutlabel="nous avons le choix dans la date\n",
-		decal=0) {
-	  hwinhor=function(h,...){
-		object=setref_horodate(object,svalue(horodate))
-		assign(nomassign,object@horodate,envir_stacomi)
-		funout(funoutlabel)
-		#print(object)
-		#dispose(winpa)
-	  }
-	  if (decal!=0){
-		# Returns the first horodate of a year shifted by decal
-		# @param horodate The horodate to shift (class POSIXt)
-		# @param decal number of year to shift
-		# @return A POSIXt
-		shiftyear<-function(horodate,decal){
-		  anneeprec=as.numeric(strftime(horodate,"%Y"))+decal
[…]
