+++ R/report_mig_char.R
@@ -1,596 +1,0 @@
-#' Migration report along with quantitative and
-#' qualitative characteristics
-#' 
-#' Migration along with qualitative or quantitative characteristics or both
-#' (e.g.) weight of eels according to the size class per period of time, weight
-#' of fish according to gender, number of fish per age class. This class does not split migration evenly over 
-#' time period. So, unlike calculations made in class report_mig and report_mig_mult
-#' the whole time span of the migration operation is not considered, only  the date of beginning of 
-#' the operation is used to perform calculations. 
-#' 
-#' @include ref_parquan.R
-#' @include ref_parqual.R
-#' @include ref_choice.R
-#' @include report_sample_char.R
-#' @note The main difference between this class and \link{report_sample_char-class} is that this class allows to
-#' select (or not) the samples, and that it handles quantitative and qualitative parameters separately.
-#' @section Objects from the Class: Objects can be created by calls of the form
-#' \code{new("report_mig_char", ...)}.  they are loaded by the interface
-#' using interface_report_mig_char function.
-#' @slot calcdata A "list" of calculated data, filled in by the calcule method
-#' @slot data A \code{data.frame} inherited from \link{report_sample_char-class}
-#' @slot dc An object of class \link{ref_dc-class} inherited from \link{report_sample_char-class}
-#' @slot taxa An object of class \link{ref_taxa-class} inherited from \link{report_sample_char-class}
-#' @slot stage An object of class \link{ref_stage-class} inherited from \link{report_sample_char-class}
-#' @slot horodatedebut An object of class \link{ref_horodate-class} inherited from \link{report_sample_char-class}
-#' @slot horodatefin An object of class \link{ref_horodate-class} inherited from \link{report_sample_char-class}
-#' @slot par An object of class \link{ref_par-class} inherited from \link{report_sample_char-class}
-#' @slot echantillon An object of class \link{ref_choice-class}, vector of choice
-#' @slot parquan An object of class \link{ref_parquan-class}, quantitative parameter 
-#' @slot parqual An object of class \link{ref_parqual-class}, qualitative parameter
-#' @family report Objects
-#' @aliases report_mig_char
-#' @author Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
-#' @concept report Object 
-#' @example inst/examples/report_mig_char-example.R
-#' @keywords classes
-#' @export
-setClass(Class="report_mig_char",
-	representation=representation(
-		echantillon="ref_choice",
-		calcdata="list",
-		parqual="ref_parqual",
-		parquan="ref_parquan"),
-	prototype=list(
-		data=list(),
-		echantillon=new("ref_choice",
-            "listechoice"=gettext(c("with","without"),domain="stacomiR"),
-			selectedvalue=gettext("with",domain="stacomiR")),
-		calcdata<-list(),
-		parqual=new("ref_parqual"),
-		parquan=new("ref_parquan")),
-	contains="report_sample_char")
-
-
-setValidity("report_mig_char",function(object)
-	{
-	  retValue=""
-	  rep4<-length(object@taxa)==1
[…]
