### 2019-02-28 #791 @MarionLegrandLogrami
Hi @boshek and @labarba - We are happy to inform you that we completed the development of the automated installation of our package stacomiR and that we commited the scripts on github. We produced two versions of the automated installation: one for Windows and the other for Linux. To date, we thoroughly tested the installation under Windows. We still have some checks to perform on the Linux version. Indeed, we checked that the script creates the database, restores the data available on github and creates an ODBC link but we still have to test the launch of the package (in particular, check that the connection is well established with the database). We plan to do these final checks (for the Linux version) by the end of next week. We hope that this development will meet your expectations. We believe that it will significantly simplify the installation of our package and make it more easily accessible. We wanted to thank @boshek for this hint.
