### 2019-03-06 619ed5a
Linux - enhanced odbc link creation
Linux install:
 - ODBC link creation in /etc/odbc.ini
 - automatically update pg_hba.conf to allow odbc user
 - better errors detections
