+++ Installation/install_stacomi.sh
@@ -2,7 +2,7 @@
 #
 # Install script for Stacomi database on Linux
 # (c)2019 LOGRAMI      - Marion LEGRAND
-# (c)2019 EPTB Vilaine - Cédric BRIAND
+# (c)2019 EPTB Vilaine - Cédric BRILLANT
 # 
 # Return codes:
 #	 0   no error
@@ -17,12 +17,15 @@
 _ROOT=$(cd $(dirname "${BASH_SOURCE[0]}") && pwd)
 ZIP_ARCH_NAME=install_bd_contmig_nat
 ZIP_ARCH=data/${ZIP_ARCH_NAME}.zip
+ODBC_LINK=data/odbc.ini
 
 # Default variables
 DB_NAME=bd_contmig_nat
 SQL_SUDO=postgres
 SQL_USER=postgres
 SQL_PASS=postgres
+SQL_IAV_USER=iav
+SQL_IAV_PASS=iav
 SQL_HOST=localhost
 SQL_PORT=5432
 REMOTE=0
@@ -60,8 +63,8 @@ PACKAGE_MANAGER=
 # Test dependency
 #	$1	program to test (empty = search for installed package name)
 #	$1	Description
-#	$2	Debian package name
-#	$3	RedHat package name
+#	$2	Debian package name (regexp)
+#	$3	RedHat package name (regexp)
 #---------------------------------------------------------
 test_dependency() {
 	local prog=$1
@@ -81,9 +84,9 @@ test_dependency() {
 	# (/!\ It could have been compiled by hand, and so not being detected here...)
 	else
 		if [ "$OS" = "debian" ]; then
-			found=$(dpkg-query -W $package 2>/dev/null)
+			found=$(dpkg -l | grep -E "^ii\s+$package\s" 2>/dev/null | awk '{print $2}')
 		elif [ "$OS" = "redhat" ]; then
-			found=$(rpm -qa | grep "^${package}-" 2>/dev/null)
+			found=$(rpm -qa | grep -E "^${package}-[0-9]" 2>/dev/null)
 		fi
 	fi
 	# found?
@@ -109,6 +112,7 @@ test_dependency() {
 			exit $RET_MISS
 		fi
 		# Install automatically
+		package=$(echo "$package" | sed -r 's~\.\*~~')
 		if [ "$OS" = "debian" ]; then
 			# The 1st time, update packages list
 			if [ $OS_UPD -eq 0 ]; then
@@ -121,12 +125,12 @@ test_dependency() {
 			result=$?
 		elif [ "$OS" = "redhat" ]; then
 			# The 1st time, update packages list
-			if [ $OS_UPD -eq 0 ]; then
-				echo -e "              [${BLUE}*${WHITE}] Doing 'yum upgrade'"
-				yum upgrade 1>/dev/null
-				OS_UPD=1
-			fi
-			echo -e "              [${BLUE}*${WHITE}] Doing 'apt-get -y install $package'"
+			#if [ $OS_UPD -eq 0 ]; then
+			#	echo -e "              [${BLUE}*${WHITE}] Doing 'yum upgrade'"
+			#	yum upgrade 1>/dev/null
+			#	OS_UPD=1
+			#fi
+			echo -e "              [${BLUE}*${WHITE}] Doing 'yum -y install $package'"
 			yum -y install $package 1>/dev/null
 			result=$?
 		fi
@@ -202,7 +206,6 @@ fi
 [ -z "$(whoami | grep '^root$')" ] && usage "This script must be executed as root. Please try:\n${BOLD}       su - root $_ROOT/$0${WHITE}"
 
 
-
 # Check dependencies
 #-------------------
[…]
