### 2019-03-06 61b576e
Linux - odbc.ini template
Template for creating ODBC link on Linux
