+++ Installation/data/odbc.ini
@@ -1,0 +1,17 @@
+
+[bd_contmig_nat]
+Description         = PostgreSQL Unicode - StacomiR
+Driver              = psqlodbcw.so
+Trace               = No
+TraceFile           = /tmp/psqlodbcstacomi.log
+Database            = bd_contmig_nat
+Servername          = localhost
+UserName            = iav
+Password            = iav
+Port                =
+ReadOnly            = Yes
+RowVersioning       = No
+ShowSystemTables    = No
+ShowOidColumn       = No
+FakeOidIndex        = No
+ConnSettings        =
