### 2019-03-06 b9af20a
Updated doc on Linux odbc creation
 - Explaining how to create ODBC link on Linux (Debian & Redhat like)
 - Fix typo
 - PostgreSQL support enhanced to 9.2 instead of 9.5
