### 2019-03-07 #791 @MarionLegrandLogrami
Hi @boshek - a little word just to let you know that we commited on github the final version of the automated installation for Linux. I forgot to tell you last time that we first tried to developp an R script to make all the installation (we tested the functions you suggested), but we failed. It's the reason we thought of an other way with batch files. We hope this will meet your expectations.
