### 2019-04-12 #791 @boshek
Hello @MarionLegrandLogrami and @arfon

It is possible that I am out of depth here. I acknowledge that. However if I am asked to provide an honest assessment here I don't think that the authors have addressed my concerns above. I am having  trouble getting the database set up on Windows. My original recommendation is that all the database set up be moved into R functions; that is wrap the shell script in an R function. In the absence of functions that do this the installation instruction need to be comprehensive - that is step by step. I am making the assumption here that the audience of this package is an average R user without any (or maybe only a little) knowledge of PostgreSQL. With that in mind this section:

https://github.com/MarionLegrandLogrami/stacomiR/tree/master/Installation#22-on-windows-host-

assumes too much knowledge on the users part.

I am unable to call the `install_stacomi.cmd`. Perhaps there is additional setup, which may be obvious, but if I am the naive new user I don't think that I would be the only person to run into these issues. For example this line:

> You WILL have to give a SQL user with admin rights (create a database, create a connection role). Else the install will fail.

assumes that a user even knows what this means or how to go about doing it. My assessment here might be reflective of my relatively poor understanding of PostgreSQL. However if the authors have a different audience for the package in mind, rather than an R user with limited knowledge of databases, perhaps this is fine. I wish to be clear in saying that this is not an indictment of the authors work. The skills and technical expertise is clear and impressive. However I am still not able to get this up an running using the current instructions.
