### 2019-04-12 #791 @arfon
Thanks for your honest feedback here @boshek - it's much appreciated.

@MarionLegrandLogrami - both of the reviewers of this submission have had serious challenges getting the database up and running for testing. While @boshek is being very gracious stating that they don't feel like an expert with these technologies I think their point that this is going to be very hard for _any_ users of your software is a very good one.

I'm not exactly sure what to suggest at this point but it's clear that the installation documentation is still lacking for new users. Perhaps you could find someone local to your host institution to try and follow the installation instructions so you can get a better understanding of what challenges people are running into.

As things stand, both reviewers have dedicated significant time and effort to these reviews but thus far, it's not been possible for either of our reviewers to fully verify the functionality of this software which means that this review is currently blocked.
