### 2019-04-19 6029300
new method to execute cmd file using shell function in R + some more details explanation on how to use the install_stacomi.cmd file
