### 2019-04-23 0bdf6b6
Modification to allow to launch cmd inside R + some clarifications in the documentation
