### 2019-04-23 #791 @MarionLegrandLogrami
Hi @boshek and @arfon,

The progam we provided to help to install everything is a cmd file. So it is an executable file. To launch it, you have to double-clik on the file. Doing this you will launch the program with some default settings. If you want to change these default settings you have to call the program in a console (start menu / cmd+enter to call the command) and enter the settings you need. I made some changes to the documentation in github to explain more clearly how to do this.
If you really want to execute this program inside R, you can call the shell function. I explained everything in the documentation:

#Inside R, if you want to launch the program with the default settings
scriptAddr<-"path_to_install_stacomi.cmd" (e.g. "C:/Users/Public/stacomir/Installation/install_stacomi.cmd")
shell(scriptAddr,"cmd")

#If you need to change some settings, you first have to look at the options provided by the install_stacomi.cmd program. To do this type:
scriptAddr<-"path_to_install_stacomi.cmd" (e.g. "C:/Users/Public/stacomir/Installation/install_stacomi.cmd")
command<-paste(scriptAddr,"-h")
shell(command,"cmd")

##To change some options:
scriptAddr<-"path_to_install_stacomi.cmd" (e.g. "C:/Users/Public/stacomir/Installation/install_stacomi.cmd")
command<-paste(scriptAddr,"-H hostname -P port -u user -p password")
#Example with hostname=localhost (=default settings), port=6543, user=foo, password=foofoo
command<-paste(scriptAddr,"-P 6543 -u foo -p foofoo") #I don't need to add -H localhost as it is the default settings
shell(command,"cmd")

I tried to clarify in the documentation the admin rights for the SQL user. If you made the postgreSQL installation yourself it is the username/password you provided during the installation. If you didn't install the program by yourself you need to ask the person who did it to give you a username with the rights to create a database and a connection role. Indeed to make graphs with the stacomiR using the graphical interface provided by the package you need to have data in a database. This database will be created and fill with a sample set of data automatically  with the cmd program but to do this you really have to have a username with the right to create a database. The user you create during the installation of PostgreSQL have all the rights needeed.

The difficulty with the installation is that we have to connect R with other programs like postgreSQL. In France and Spain, several structures use this program and as these users often have no knowledge neither in R nor in database, we provide a support of help to all the users by realizing ourselves the installation of everything that is needed on their computers.
Thanks to our wish to publish our package in JOSS and the comments of our 2 reviewers, we have done a lot of work to simplify the installation. But actually, the user must always have some knowledge including how PostgreSQL works if he wants to be autonomous in the installation (otherwise, he must contact us so we can help). We can hardly see how to do otherwise. Indeed, if the user does not know if postgreSQL is installed on his computer on a localhost or on a distant server or does not have in mind the user and/or the password provided during the installation of postgres this can cause difficulties during installation. In our cmd program we have entered default values ​​for the server (= localhost), for the port (= 5432 i.e. the default port on which postgresql is installed) and for user and password (postgres for both). But the user will certainly have to change at least the user and the password (except if he chose this user and password during the installation). This is possible either with the command prompt or now in R with the shell function.
We hope that all these informations will help.
@boshek: don't hesitate to contact us if there are still things that are not clear. We will be happy to help you. We spent lot of time trying to improve our package and we really hope to publish it in JOSS
