### 2019-05-02 #791 @boshek
Hi @MarionLegrandLogrami - just a quick check in here to ask if you've considered putting all this into a docker container which can reproducibly run. That was you'd only have to provide instructions to the user on how to use docker. How feasible do you think that would be?
