### 2019-05-03 #791 @MarionLegrandLogrami
Hi @boshek - Thank you for your message. We spent some time reading a little about the docker container with which we are not familiar. This technology seems very interesting, but we are not sure that it will facilitate the installation (as it seem to be not really easy to use docker especially if the goal is to run an R package with a graphical interface - how can you access the graphical interface and interact with it inside the docker is still a mistery to me). What about the shell() command we suggested to use to launch the installation inside R ? Have you had any difficulties ? Have you managed to install postgresql ? If you encounter difficulties with any part of the installation don't hesitate to contact us. We will be very happy to help you (for example by using Teamviewer to help you to install everything)
