### 2019-05-11 #791 @boshek
Hi @MarionLegrandLogrami - I still think that the whole script can be abstracted away into an R function and the script itself could be included in the package in `inst` folder. Arguments to think function could be the username, password, host and port. That function would run the script for you. I think if you asked users if they were ok with the function running an executable you'd be ok.

In terms of the process, I am having trouble with the odbc part of the installation. On my machine the 64 bit odbc driver causes a silent crash of pgAdmin and no erro message. Installation of the 32 bit driver proceeds but the program executable seems unable to find the driver. Certainly the shell script you have written is very useful. However I am still stuck. Sorry.
