### 2019-05-13 #791 @MarionLegrandLogrami
Hi @boshek - thank you for this detailed feedback on the problems encountered during installation. Depending on the error message, the problem is that you have installed an odbc driver that does not rely on the same architecture than the one used by R (you have a 64-bit R version and a 32-bit ODBC driver). There are 2 solutions:
1. Install the 64 bit odbc driver. The best way is to take the latest version of the driver regardless of the version of postgresql (I modified this instruction in the documentation). You can [download it here](https://ftp.postgresql.org/pub/odbc/versions/msi/psqlodbc_11_00_0000-x64.zip)
2. open 32-bit R and keep your 32-bit odbc driver

Regarding your first suggestion (include everything in the package), we unfortunately can not follow it because for the package to be accepted by CRAN, it is necessary that the data does not exceed 5MB.

The compressed database is already 8 MB. If we try to include it in the package, we may not be able to submit it to CRAN.
![image](https://user-images.githubusercontent.com/26621584/57627718-d0331c00-7598-11e9-88dd-e4a48957d213.png)
(read on the [CRAN Repository Policy pages](https://cran.r-project.org/web/packages/policies.html))

Your feedback on the quality of the documentation seems very important for us. Unfortunately, I am not sure I understand the suggestion you made to improve the documentation. At this point, I have separated the odbc driver installation detail in a specific part of the documentation (part 6. More help), leaving only in the "Requirement" part the indication that you need to have an odbc driver and indicating a link to the specific part of the document in which I explain how to make this installation. You propose to separate this part. You mean to extract it completely from the documentation to put it in another help file? If you could give me some details I will be very grateful
