### 2019-06-05 #791 @MarionLegrandLogrami
Hello @boshek - Did you manage to install the 64-bit postgresql driver or did you try the 32-bit R version compatible with the postgresql driver that you already installed?
Please, let me know if you have any problem !
