### 2019-07-22 #791 @MarionLegrandLogrami
@kyleniemeyer - Dear Editor, we are very grateful for the time spent by our reviewers on the submission of our package to JOSS and for their comments. Thanks to them we made lot of progress ! I was planning to use this paper as part of my Phd Thesis for which time is running out as we are waiting to see our package accepted by JOSS. This is a package that has passed all R CRAN checks and is already used by many organizations on a daily basis, and we are freely and actively advising and helping end users, mostly people doing fish migration monitoring in the field. To be fully functional, the package requires a database connection which makes the installation more complex. We offered our reviewer to help him install the various necessary programs and also developed a specific module to help him install postgres and set up the connection. To date we have trouble understanding where he got troubles in the installation. So we are coming back to you to see if it could be possible to speed up the validation process of our package in JOSS ? Again we are at your disposal and that of the reviewer to do anything to help you. Thanking you in advance for your help.
