### 2019-08-01 #791 @boshek
Ok I will commit to having this finished by the end of August 9th. Sorry to have languished so long here. I still think that too much complexity is being offloaded to the user here rather than abstracting that away.
