### 2019-08-10 #791 @boshek
This is an update to this process.

## Using the stacomi package

I was running into issues because of leftover installations and other uses of a local PostgreSQL server. Additionally I was running into issues because I had installed the ODBC driver via the StackBuilder. That installed the driver in a different place than manually installing like the instructions said. **I would recommend having your installation script check both locations for the driver.**

That said I have been able to only go a little further in the process in getting the package/database up and running. I am able to successfully bring up the UI using `stacomi()`. I do get this error however:

```

Loading required package: stacomirtools
Loading required package: RODBC
> stacomi()
C:/program files/stacomi/calcmig.csv does not exist, switching to defaut package fileLoading required package: gWidgetsRGtk2
Loading required package: RGtk2
Loading required package: gWidgets
Loading required package: cairoDevice
```
That said everything proceed in the sense that the dashboard appears. However I am not able to make it do anything (generate figures or tables) and I am assuming that the connection to the database isn't functioning properly. My first thought was to manually connect to the database in R and check if any actual data is there:

``` r
library(DBI)
library(dplyr)
#>
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#>
#>     filter, lag
#> The following objects are masked from 'package:base':
#>
#>     intersect, setdiff, setequal, union
library(dbplyr)
#>
#> Attaching package: 'dbplyr'
#> The following objects are masked from 'package:dplyr':
#>
#>     ident, sql

con <- dbConnect(RPostgres::Postgres(),dbname = 'bd_contmig_nat',
                 host = 'localhost',
                 port = 5432,
                 user = 'postgres',
                 password = 'foo')

dbListTables(con)
#> [1] "chevauchement"

tbl(con, in_schema("iav", "t_bilanmigrationjournalier_bjo_bjo_identifiant_seq"))
#> # Source:   table<iav.t_bilanmigrationjournalier_bjo_bjo_identifiant_seq>
#> #   [?? x 3]
#> # Database: postgres [postgres@localhost:5432/bd_contmig_nat]
#>   last_value log_cnt is_called
#>   <int64>    <int64> <lgl>
#> 1 351677     0       TRUE
```

I am unclear that this output is telling me and whether it is helpful at all. However at this point this is as far as I could get. I thought to try installing the dev version but got some GTK errors which I am unable to debug at this point:

> Warning: Failed to load RGtk2 dynamic library, attempting to install it.
Please install GTK+ from http://ftp.gnome.org/pub/gnome/binaries/win32/gtk+/2.22/gtk+-bundle_2.22.1-20101227_win32.zip
If the package still does not load, please ensure that GTK+ is installed and that it is on your PATH environment variable
IN ANY CASE, RESTART R BEFORE TRYING TO LOAD THE PACKAGE AGAIN
Error in inDL(x, as.logical(local), as.logical(now), ...) :
  unable to load shared object 'C:/Users/samal/Documents/R/win-library/3.6/RGtk2/libs/i386/RGtk2.dll':
  LoadLibrary failure:  The specified module could not be found.

## Some Other Considerations

### Installation Script
I still think that the installation script could be written completely in R avoiding the clunky step of downloading the installation file and running it. I think that you have misunderstood my meaning here. I don't want to run the installation script with R. I think it should be written in R itself so that a user can debug in the language they are most likely to be familiar with - R. This would provide a unified interface for users. However, I think given the length of this process and admitting that reasonable people can disagree on this point, this is not a point that is worth holding up acceptance of this review particularly since my more material concerns have been ameliorated by the process that the installation script enacts and very thorough documentation.

### Shiny
After finally being able to generate the UI for this process I can't help but ask why you aren't doing this in Shiny. This comment is only relevant to this review because it would great ease a user experience (mine being documented here) taking the target audience for a dashboard (non R users) and taking them completely out of R. Without even considering the much improved UI experience, Shiny would be far easier for you to deploy. For a future I would strongly recommend using Shiny for the functionality currently being filled for GTK+.
