### 2019-08-15 #791 @arfon
This review has been going on for a very long time now. I'm very grateful to both @steffilazerte and @boshek for their work here - I believe the `stacomiR` package is in substantially better shape because of their time and efforts.

While I believe there are still some challenges with this software, I'm going to make the editorial decision at this point to accept this submission into JOSS.
