### 2019-08-19 b20c6a3
Package version as accepted by JOSS
