### 2019-08-19 #791 @MarionLegrandLogrami
Thank you very much for accepting this submission in JOSS and @steffilazerte and @boshek - thanks again for your time and advice!

@arfon - Here's the DOI (for the archive of the package):
[http://doi.org/10.5281/zenodo.3371259](http://doi.org/10.5281/zenodo.3371259)
